pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.5.2",
    "google-auth>=2.40.3",
    "gspread>=6.2.1",
    "numpy>=2.3.1",
    "oauth2client>=4.1.3",
    "pandas>=2.3.1",
    "requests>=2.32.4",
]
